requirements.txt
amqp
asgiref
attrs
autobahn
Automat
billiard
celery
certifi
cffi
channels
channels_redis
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
constantly
cryptography
daphne
dj-rest-auth
Django
django-allauth
django-appconf
django-cors-headers
django-extensions
django-filter
django-jazzmin
django-phonenumber-field
django-shortuuidfield
djangorestframework
djangorestframework_simplejwt
graphviz
hyperlink
idna
incremental
kombu
msgpack
paho-mqtt
phonenumbers
pillow
prompt_toolkit
pyasn1
pyasn1_modules
pycparser
pydotplus
PyJWT
PyMySQL
pyOpenSSL
pyparsing
python-dateutil
python-decouple
redis
requests
service-identity
setuptools
shortuuid
six
sqlparse
Twisted
txaio
typing_extensions
tzdata
urllib3
vine
wcwidth
whitenoise
zope.interface
